backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
faster-whisper==1.0.0
twilio==9.2.3
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.26
soundfile==0.12.1
python-dotenv==1.0.0
SpeechRecognition==3.10.0
PyAudio==0.2.14
